Refactor network communication; streamline NetworkManager and ClientService methods
